Restructure components, cleanup
Changes:
1. Lazy load home and movie details
2. Added home placeholder as suspense for movie list (home)
3. Movie list a functional component
4. Added movie placeholder as suspense for movie details
5. Handled 'Offline' status for all routes
6. Added Error Boundaries
