+++ src/App.js
@@ -1,23 +1,68 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-import Header from "./header/header";
-import Main from './main/Main';
+import { isOffline } from './components/selectors/is-offline';
+import Home from './components/Home';
+import Header from './components/header';
+import MovieDetails from './components/Movie';
 import './App.css';
 
-const LazyDetails = React.lazy(() => import('./main/movies/Movie'));
-
 class App extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      hasError: false,
+      offline: isOffline(),
+    }
+  }
+
+  getDerivedStateFromProps(props) {
+    return {
+      offline: isOffline(),
+    }
+  }
+
+  updateConnection = status => {
+    console.log(`>> updating offline status to ${status}`);
+    this.setState({offline: status})
+  }
+
+  componentDidMount() {
+    document.addEventListener('offline', () => this.updateConnection(true), false);
+    document.addEventListener('online', () => this.updateConnection(false), false);
+  }
+
+  componentWillUnmount() {
+      document.removeEventListener('offline', () => this.updateConnection(null), false);
+      document.removeEventListener('online', () => this.updateConnection(false), false);
+  }
+
   render() {
+    const { offline } = this.state;
+
     return (
         <BrowserRouter>
           <div className="App">
             <Header key='moview-header'/>
-              <React.Suspense fallback={<div> Loading ... </div>}>
-                <Switch>
-                    <Route exact path="/" component={Main} />
-                    <Route path="/movies/:movieId" component={LazyDetails} />
-                </Switch>
-              </React.Suspense>
+            <Switch>
+                <Route exact
+                  path="/"
+                  render={(props) => 
+                    <Home
+                      offline={offline}
+                      {...props} 
+                    />
+                  }
+                />
+                <Route
+                  path="/movies/:movieId"
+                  render={(props) =>
+                    <MovieDetails
+                      offline={offline}
+                      {...props}
+                    />
+                  }
+                />
+            </Switch>
           </div>
         </BrowserRouter>
     );
