+++ src/components/Home/index.jsx
@@ -1,0 +1,12 @@
+import React, { lazy, Suspense } from 'react';
+import MovieListLoader from './placeholder';
+
+const MovieHome = lazy(() => import('./movies'));
+
+const MoviesList = (props) => (
+    <Suspense fallback={<MovieListLoader />}>
+        <MovieHome {...props} />
+    </Suspense>
+);
+
+export default MoviesList;
