+++ src/components/Home/navigation/Navigation.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { KEY_CODES, MOVIE_GENRES } from '../../constants';
+import { KEY_CODES, MOVIE_GENRES } from '../../../constants';
 import Selection from "./Selection";
 import Slider from './Slider';
 import SearchButton from './SearchButton'
