+++ src/components/Home/navigation/Slider.js
