+++ src/components/Movie/index.jsx
@@ -1,0 +1,12 @@
+import React, { lazy, Suspense } from 'react';
+import MovieListLoader from './placeholder';
+
+const Movie = lazy(() => import('./movie'));
+
+const MovieItem = (props) => (
+    <Suspense fallback={<MovieListLoader />}>
+        <Movie {...props} />
+    </Suspense>
+);
+
+export default MovieItem;
