+++ src/components/Movie/styles.css
@@ -72,7 +72,33 @@
 }
 
 .movie-card {
+    align-items: baseline;
     display: flex;
     justify-content: space-between !important;
-    align-items: baseline;
+    margin-top: 0.5rem !important;
+}
+
+/*******************************
+People component styles
+*******************************/
+.ph-card {
+    width: inherit !important;
+    max-width: 100%;
+}
+
+.ph-image {
+    max-width: 100% !important;
+    max-height: 24rem !important;
+}
+
+.ph-details {
+    max-width: inherit !important;
+}
+
+.credits {
+    margin: 1rem 0rem;
+}
+
+.error-image {
+    width: 50% !important;
 }
