+++ src/components/common/index.js
@@ -1,0 +1,13 @@
+import ErrorPage from './error-page';
+import Form from './form';
+import LazyImage from './backdrop-image';
+import List from './list';
+import MovieRating from './rating';
+import NoConnection from './no-connection';
+
+export { ErrorPage };
+export { Form as SearchMovie };
+export { LazyImage as Image };
+export { List };
+export { MovieRating as Rating };
+export { NoConnection };
