+++ src/components/common/no-connection.jsx
@@ -1,0 +1,11 @@
+import React, { PureComponent } from 'react'
+
+export default class NoConnection extends PureComponent {
+  render() {
+    return (
+      <div>
+        <h4> You're not connected to the Internet </h4>
+      </div>
+    )
+  }
+}
