+++ src/components/common/rating.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './styles.css';
 
 const DEFAULT_RATING = 3;
 const MAX_INPUT_RATING = 10;
@@ -13,14 +14,21 @@ const MovieRatings = ({ title, vote_average = DEFAULT_RATING }) => {
 
     for (let i=0; i < Math.floor(fiveStars); i++) {
         let key = getRatingKey(title, fiveStars + i);
-        fullStars.push(<i key={key} class="fas fa-star fa-xs"></i>);
+        fullStars.push(<i key={key} class="fas fa-star fa-xs star"></i>);
     }
 
     if (fraction) {
-        fullStars.push(<i key={getRatingKey(title, 2*fiveStars)} class="fas fa-star-half fa-xs"></i>)
+        fullStars.push(<i key={getRatingKey(title, 2*fiveStars)} class="fas fa-star-half fa-xs star"></i>)
     }
 
-    return <div style={{color: 'orange'}}>{ fullStars }</div>;
+    return (
+        <div
+            style={{
+                color: 'orange'
+            }}>
+            { fullStars }
+        </div>
+    );
 };
 
 export default MovieRatings;
