+++ src/components/common/styles.css
@@ -1,0 +1,130 @@
+.movie-backdrop {
+    background-size: cover;
+    background-position: center center;
+}
+
+/*******************************
+FORM styles
+*******************************/
+form {
+  position: relative;
+  padding: 0.5rem;
+  text-align: center;
+  margin: 0px !important;
+}
+
+.searchBar {
+  position: relative;
+  font-family: 'Open Sans', sans-serif;
+  width: 100%;
+  padding: 20px 0 20px 40px;
+  border: 1px solid #bdbdbd;
+  color: #000;
+  outline: none;
+  transition: all 0.35s ease-out;
+}
+
+.searchBar:focus {
+  border: 1px solid #1DE9B6;
+}
+
+::placeholder {
+  color: #000;
+  transition: all 0.35s ease-out;
+}
+
+.searchBar:focus::placeholder {
+  color: #1DE9B6;
+}
+
+.searchField {
+  margin: 0px !important;
+}
+
+.searchIcon {
+  position: absolute;
+  z-index: 9999;
+  margin: 4px;
+  width: 100% !important;
+}
+
+.ui.form .field .ui.input input, .ui.form .fields .field .ui.input input {
+  width: 100% !important;
+}
+
+.noDisplay {
+  display: none;
+  opacity: 0;
+  visibility: hidden;
+}
+
+/*******************************
+FORM results styles
+*******************************/
+
+.formResults {
+  position: absolute;
+  z-index: 9999;
+  width: 100%;
+  height: 500px;
+  overflow-y: visible;
+  text-align: left;
+  padding: 0px !important;
+  margin: 0px !important;
+}
+
+.formResults li, .sitem {
+  background-color: rgb(100, 100, 241);
+  transition: all 0.35s ease-out;
+}
+
+.formResults li:hover {
+  background-color:rgb(21, 21, 114);
+}
+
+.formResults a {
+  color: #fff;
+  display: flex;
+  padding: 8px !important;
+}
+
+.resultPoster {
+  width: 25%;
+  margin-right: 16px;
+  align-self: flex-start;
+}
+
+/*******************************************
+  LIST ITEM styles
+*******************************************/
+
+.credits-label {
+  background-color: inherit !important;
+  border: none !important;
+}
+
+.credits-label-sec {
+  background-color: inherit !important;
+  border: none !important;
+  font-weight: 100 !important;
+}
+
+.average {
+  color: rgba(226,122,63, 0.9);
+  display: flex !important;
+  flex-basis: 40%;
+  flex-direction: row;
+  flex-wrap: wrap;
+  font-size: 14px;
+  justify-content: center;
+  padding: 4px !important;
+}
+
+/*******************************************
+    RATING styles
+*******************************************/
+
+.star {
+  padding: 2px !important;
+  line-height: 2.25;
+}
