--- src/components/form.css
@@ -1,51 +1,0 @@
-form {
-  position: relative;
-  padding: 0.5rem;
-  text-align: center;
-  margin: 0px !important;
-}
-
-.searchBar {
-  position: relative;
-  font-family: 'Open Sans', sans-serif;
-  width: 100%;
-  padding: 20px 0 20px 40px;
-  border: 1px solid #bdbdbd;
-  color: #000;
-  outline: none;
-  transition: all 0.35s ease-out;
-}
-
-.searchBar:focus {
-  border: 1px solid #1DE9B6;
-}
-
-::placeholder {
-  color: #000;
-  transition: all 0.35s ease-out;
-}
-
-.searchBar:focus::placeholder {
-  color: #1DE9B6;
-}
-
-.searchField {
-  margin: 0px !important;
-}
-
-.searchIcon {
-  position: absolute;
-  z-index: 9999;
-  margin: 4px;
-  width: 100% !important;
-}
-
-.ui.form .field .ui.input input, .ui.form .fields .field .ui.input input {
-  width: 100% !important;
-}
-
-.noDisplay {
-  display: none;
-  opacity: 0;
-  visibility: hidden;
-}
