--- src/components/formResults.css
@@ -1,31 +1,0 @@
-.formResults {
-  position: absolute;
-  z-index: 9999;
-  width: 100%;
-  height: 500px;
-  overflow-y: visible;
-  text-align: left;
-  padding: 0px !important;
-  margin: 0px !important;
-}
-
-.formResults li, .sitem {
-  background-color: rgb(100, 100, 241);
-  transition: all 0.35s ease-out;
-}
-
-.formResults li:hover {
-  background-color:rgb(21, 21, 114);
-}
-
-.formResults a {
-  color: #fff;
-  display: flex;
-  padding: 8px !important;
-}
-
-.resultPoster {
-  width: 25%;
-  margin-right: 16px;
-  align-self: flex-start;
-}
