--- src/components/index.js
@@ -1,9 +1,0 @@
-import BackdropImage from './backdrop-image';
-import List from './list-item';
-import Rating from './rating';
-
-export {
-  BackdropImage as LazyLoad,
-  List as ListItem,
-  Rating,
-};
