+++ src/components/selectors/is-offline.js
@@ -1,0 +1,12 @@
+
+export const isOffline = () => {
+    if (!!navigator && navigator.onLine) {
+        return !navigator.onLine;
+    }
+
+    if (window.navigator && window.navigator.onLine) {
+        return !window.navigator.onLine;
+    }
+
+    return true;
+}
