+++ src/components/styles.css
@@ -1,8 +1,3 @@
-.movie-backdrop {
-    background-size: cover;
-    background-position: center center;
-}
-
 .movie-poster-img {
     min-height: 128px;
     margin-top: 8px;
