--- src/containers/index.js
@@ -1,12 +1,0 @@
-import ErrorPage from './error-page';
-import List from "./list";
-import Credits from './credits';
-import PlaceHolder from './home-list-placeholder';
-import DetailsHolder from './movie-details-placeholder';
-
-export { Credits };
-export { ErrorPage };
-export { List };
-export { PlaceHolder as MovieListLoader };
-export { DetailsHolder as MovieDetailsHolder };
-
