--- src/main/Main.css
@@ -1,107 +1,0 @@
-.main {
-    display: flex;
-    flex-direction: column;
-}
-
-.main .filter {
-    color: blue;
-    position: fixed;
-    bottom: 2rem;
-    right: 2rem;
-    z-index: 3;
-    border: 2px solid #e27a3f;
-    padding: 8px;
-}
-
-
-.modal-custom {
-    border-radius: 4px;
-}
-
-.search-label {
-    color: #3f51b5;
-    font-size: 15px;
-}
-
-.now-showing {
-    font-weight: bolder;
-    border-bottom: 4px solid orangered;
-    max-width: 100px;
-    margin: 8px;
-    padding: 4px 0px 4px 0px;
-}
-
-.list-info {
-    font-size: 12px;
-    flex: 1;
-    display: flex;
-    height: 44px;
-    justify-content: space-evenly;
-    align-items: center;
-    margin: 8px 4px;
-}
-
-.genre-filter {
-    border: 1px solid rgb(127, 127, 255);
-    color: rgb(255, 255, 255);
-    border-radius: 8px;
-    font-weight: bold;
-    padding: 4px;
-    margin: 0px 8px;
-    background-color: rgb(127, 127, 255);
-}
-
-.year-filter,
-.time-filter,
-.rating-filter {
-    border: 1px solid rgb(127, 127, 255);
-    background-color: rgb(127, 127, 255);
-    color: rgb(255, 255, 255);
-    font-weight: bold;
-    max-width: 96px;
-    margin: 0px 4px;
-    padding: 4px;
-    border-radius: 8px;
-    flex: 1;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.year-filter > div,
-.time-filter > div,
-.rating-filter > div {
-    margin: 4px 0px;
-}
-
-
-.year-filter > i,
-.time-filter > i,
-.rating-filter > i {
-    align-self: flex-end;
-}
-
-@media only screen and (max-width: 800px) and (max-height: 500px)  {
-    .search-label {
-        position: absolute;
-        margin-left: 80%;
-        margin-top: -7%;
-    }
-}
-
-@media only screen and (max-height: 800px) and (max-width: 500px) {
-    .search-label {
-        position: absolute;
-        margin-left: 80%;
-        margin-top: -11%;
-    }
-}
-
-@media only screen and (min-width: 800px) {
-    .search-label {
-        cursor: pointer;
-        position: absolute;
-        margin-left: 85%;
-        margin-top: -3.2%;
-    }
-}
